pyproject.toml
[project]
name = "practicumbot"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.13"
dependencies = [
    "aiogram>=3.25.0",
    "gigachat>=0.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
]
